GAIO

This project was founded by the European Region Development Fund.

GAIO.DepthFilterNodes

GAIO.Leaves

GAIO.PreOrderDFS

GAIO.PreOrderDFSFiltered

GAIO.PreOrderDFSSubIterator

GAIO.:≺

GAIO.add_sub_box!

GAIO.contains

GAIO.depth

GAIO.eval!

GAIO.init_box_tree

GAIO.push_site!

GAIO.push_site!

GAIO.push_sites!

GAIO.push_sites!

GAIO.results

GAIO.sites

GAIO.subdivide!

GAIO.subdivide!

GAIO.DepthFilterNodes — Type

DepthFilterNodes{T} <: PreOrderDFSSubIterator

An iterator based on PreOrderDFS that returns only nodes with a certain depth(s). Traversal starts at start_node but depth is measured by depth( any_node ) and hence in relation to the root of the tree (which possibly does not equal start_node) by default.

A subtree is only visited for traversal if the depth of its respective parent is below the maximum value of desired return depths.

Initialize via

dfs_iterator = DepthFilterNodes( start_node, desired_depths :: Vector{Int} );

GAIO.Leaves — Type

Leaves{T} <: PreOrderDFSSubIterator

Initialize as per iter = Leaves( start_node ) iter is an iterator that returns the leaves of a subtree starting at start_node. It is based on a PreOrderDFS traversal.

GAIO.PreOrderDFS — Type

PreOrderDFS{T}
	start_node :: T
	traverse_subtree_filter :: Union{Nothing, F where F<:Function};
end

An iterator to perform a depth first traversal of the subtree that has start_node as its root. PreOrder means that parent nodes are visited before their respective children. The field traverse_subtree_filter defaults to nothing. If a function is provided it should accept an argument of type T (a parent node) and return true if its children should be pushed to the stack for visiting.

GAIO.PreOrderDFSFiltered — Type

PreOrderDFSFiltered{T} <: PreOrderDFSSubIterator

An iterator that is based on the PreOrderDFS iterator and allows for filtering out nodes that are returned. The most basic constructor is

iter = PreOrderDFSFiltered( 
	start_node :: T, 
	func :: F where F <: Function )

The function func should take one argument of type T (a node) and return true if it should be returned by iter.

GAIO.PreOrderDFSSubIterator — Type

An abstract super type for iterators that are based on the PreOrderDFS iterator.

GAIO.:≺ — Method

≺( X, Y )

Return true if vector X dominates Y, i.e. if all components of X not bigger than those of Y and if there is at least one component of X that is strictly smaller than the corresponding Y component.

GAIO.add_sub_box! — Method

add_sub_box!( parent, sub_lb, sub_ub )

Add a new box node with variable boundaries sub_lb and sub_ub to parent. Modifies sub_boxes field of parent and sets the parent reference of the new box correctly.

GAIO.contains — Method

contains(box, point)

Return true if point lies in box.

GAIO.depth — Method

depth(node)

Return depth of node in a tree. Root node is assigned depth 0, its children have depth 1 and so forth.

This default implementation goas up the tree starting at node as long as there is a parent node.

It can be overwritten for custom Node/Tree types, e.g. when depth is stored explicitly.

GAIO.eval! — Method

Evaluate all sites referenced by field db_indices.

GAIO.init_box_tree — Function

init_box_tree( lb, ub, db = nothing )

Initialize the root node of a box Tree with variable boundaries lb and ub. If db is a database then it is referenced in the node.

GAIO.push_site! — Function

Push a single site into db, set Evaluation result to nothing.

GAIO.push_site! — Function

Push a single site into db, set Evaluation result to nothing.

GAIO.push_sites! — Method

Push multiple sites into db, set Evaluation results to nothing.

GAIO.push_sites! — Method

Push multiple sites into db, set Evaluation results to nothing.

GAIO.results — Method

Retrieve evaluation results indexed by db_indices from referenced DataBase db.

GAIO.sites — Method

Retrieve sites indexed by db_indices from referenced DataBase db.

GAIO.subdivide! — Function

subdivide!(parent, divisor=2)

Recursively subdivide node parent. parent is divided by divisor along each box dimension.

GAIO.subdivide! — Method

Divide each of bns edges by divisor or the entries of divisor and add sub boxes as children to the tree.