Library Reference

GAIO.BoxMap

GAIO.Node

Arpack.eigs

Arpack.eigs

Arpack.eigs

Arpack.svds

Arpack.svds

Arpack.svds

Base.sum

GAIO.AdaptiveBoxMap

GAIO.GridBoxMap

GAIO.MonteCarloBoxMap

GAIO.PointDiscretizedBoxMap

GAIO.adaptive_newton_step

GAIO.approx_lipschitz

GAIO.armijo_rule

GAIO.bounded_point_to_key

GAIO.box_dimension

GAIO.center

GAIO.center

GAIO.chain_recurrent_set

GAIO.cover

GAIO.cover_manifold

GAIO.cover_roots

GAIO.depth

GAIO.expon

GAIO.find_at_depth

GAIO.finite_time_lyapunov_exponents

GAIO.finite_time_lyapunov_exponents

GAIO.fixqr!

GAIO.hidden_keys

GAIO.index_pair

GAIO.index_quad

GAIO.index_to_key

GAIO.key_to_box

GAIO.key_to_index

GAIO.leaves

GAIO.linreg

GAIO.marginal

GAIO.marginal

GAIO.marginal

GAIO.maximal_backward_invariant_set

GAIO.morse_adjacencies

GAIO.morse_adjacencies_and_tiles

GAIO.morse_component_map

GAIO.morse_map

GAIO.morse_tiles

GAIO.neighborhood

GAIO.nth_iterate_jacobian

GAIO.point_to_box

GAIO.point_to_key

GAIO.preimage

GAIO.preimage

GAIO.radius

GAIO.rescale

GAIO.rescale

GAIO.rescale

GAIO.rk4

GAIO.rk4_flow_map

GAIO.sample_adaptive

GAIO.seba

GAIO.seba

GAIO.subdivide

GAIO.subdivide!

GAIO.symmetric_image

GAIO.tree_search

GAIO.unstable_set

GAIO.vertices

GAIO.@save

GAIO.BoxMap — Method

BoxMap(map, domain; no_of_points=ntuple(_->4, N)) -> SampledBoxMap

Transforms a $map: Q → Q$ defined on points in the domain $Q ⊂ ℝᴺ$ to a SampledBoxMap defined on Boxes.

By default uses adaptive test-point sampling. For SIMD- and GPU-accelerated BoxMaps, uses a grid of test points by default.

GAIO.Node — Type

Node structure used for TreePartitions

Fields:

left and right refer to indices w.r.t.

trp.nodes for a TreePartition trp.

Arpack.eigs — Method

eigs(gstar::TransferOperator [; kwargs...]) -> (d,[v,],nconv,niter,nmult,resid)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Arpack.eigs — Method

eigs(gstar::TransferOperator [; kwargs...]) -> (d,[v,],nconv,niter,nmult,resid)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Arpack.eigs — Method

eigs(gstar::TransferOperator [; kwargs...]) -> (d,[v,],nconv,niter,nmult,resid)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Arpack.svds — Method

svds(gstar::TransferOperator [; kwargs...]) -> ([U,], σ, [V,], nconv, niter, nmult, resid)

Compute a set of

singular values σ

left singular vectors U

right singular vectors V

of gstar, where U and V are Vectors of BoxFuns. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.svds can be passed. See the documentation for Arpack.svds.

Arpack.svds — Method

svds(gstar::TransferOperator [; kwargs...]) -> ([U,], σ, [V,], nconv, niter, nmult, resid)

Compute a set of

singular values σ

left singular vectors U

right singular vectors V

of gstar, where U and V are Vectors of BoxFuns. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.svds can be passed. See the documentation for Arpack.svds.

Arpack.svds — Method

svds(gstar::TransferOperator [; kwargs...]) -> ([U,], σ, [V,], nconv, niter, nmult, resid)

Compute a set of

singular values σ

left singular vectors U

right singular vectors V

of gstar, where U and V are Vectors of BoxFuns. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.svds can be passed. See the documentation for Arpack.svds.

Base.sum — Method

sum(f, μ::BoxFun)
sum(f, μ::BoxFun, B::BoxSet)
μ(B) = sum(x->1, μ, B)

Integrate a function f using μ as a density, that is, if boxfun is the discretization of a measure $\mu$ over the domain $Q$, then approximate the value of

\[\int_Q f \, d\mu .\]

If a BoxSet B is passed as the third argument, then the integration is restricted to the boxes in B

\[\int_{Q \cap \bigcup_{b \in B} b} f \, d\mu .\]

The notation μ(B) is offered to compute $\mu (\bigcup_{b \in B} b)$.

GAIO.AdaptiveBoxMap — Method

BoxMap(:adaptive, f, domain::Box) -> SampledBoxMap

Construct a SampledBoxMap which uses sample_adaptive to generate test points as described in

Oliver Junge. “Rigorous discretization of subdivision techniques”. In: International Conference on Differential Equations. Ed. by B. Fiedler, K. Gröger, and J. Sprekels. 1999.

GAIO.GridBoxMap — Method

BoxMap(:grid, map, domain::Box{N}; n_points::NTuple{N} = ntuple(_->16, N)) -> SampledBoxMap

Construct a SampledBoxMap that uses a grid of test points. The size of the grid is defined by n_points, which is a tuple of length equal to the dimension of the domain.

GAIO.MonteCarloBoxMap — Method

BoxMap(:montecarlo, map, domain::Box{N}; n_points=16*N) -> SampledBoxMap

Construct a SampledBoxMap that uses n_points Monte-Carlo test points.

GAIO.PointDiscretizedBoxMap — Method

BoxMap(:pointdiscretized, map, domain, points) -> SampledBoxMap

Construct a SampledBoxMap that uses the iterator points as test points. points must be an array or iterator of test points within the unit cube [-1,1]^N.

GAIO.adaptive_newton_step — Function

adaptive_newton_step(g, g_jacobian, x, k=1)

Return one step of the adaptive Newton algorithm for the point x.

GAIO.approx_lipschitz — Method

approx_lipschitz(f, center::SVector, radius::SVector, accel=nothing) -> Matrix

Compute an approximation of the Lipschitz matrix, i.e. the matrix that satisifies

\[| f(x) - f(y) | \leq L | x - y | \quad \forall \, x,y \in \text{Box(center, radius)}\]

componentwise.

GAIO.armijo_rule — Function

armijo_rule(g, Dg, x, d, σ=1e-4, ρ=0.8, α₀=0.05, α₁=1.0)

Find a step size multiplier $\alpha \in (\alpha_0, \alpha_1]$ such that

\[g(x + \alpha d) - g(x) \leq \alpha \sigma \, Dg(x) \cdot d\]

This is done by initializing $\alpha = 1$ and testing the above condition. If it is not satisfied, scale $\alpha$ by some constant $\rho < 1$ (i.e. set $\alpha = \rho \cdot \alpha$), and test the condition again.

GAIO.bounded_point_to_key — Method

bounded_point_to_key(P::BoxPartition, point)

Find the cartesian index of the nearest box within a BoxPartition to a point. Conicides with point_to_key if the point lies in the partition. Default behavior is to set NaN = Inf if NaNs are present in point.

GAIO.box_dimension — Method

box_dimension(boxsets) -> D

For an iterator boxsets of (successively finer) BoxSets, compute the box dimension D.

Example

# F is some BoxMap, S is some BoxSet
box_dimension( relative_attractor(F, S, steps=k) for k in 1:20 )

GAIO.center — Method

center(center, radius)

Return the center of a box as an iterable. Default function for image_points in SampledBoxMaps.

GAIO.center — Method

center(b::Box)

Return the center of a box.

GAIO.chain_recurrent_set — Method

chain_recurrent_set(F::BoxMap, B::BoxSet; steps=12) -> BoxSet

Compute the chain recurrent set over the box set B. B should be a (coarse) covering of the relative attractor, e.g. B = cover(P, :) for a partition P.

GAIO.cover — Function

BoxSet constructors:

set of all boxes in partition / box set P:

julia B = cover(P, :)

cover the point x, or points x = [x_1, x_2, x_3] # etc ... using boxes from P

julia B = cover(P, x)

a covering of S using boxes from P

julia S = [Box(center_1, radius_1), Box(center_2, radius_2), Box(center_3, radius_3)] # etc... B = cover(P, S)

Return a subset of the partition or box set P based on the second argument.

GAIO.cover_manifold — Method

cover_manifold(f, B::BoxSet; steps=12)

Use interval arithmetic to compute a covering of an implicitly defined manifold $M$ of the form

\[f(M) \equiv 0\]

for some function $f : \mathbb{R}^N \to \mathbb{R}$.

The starting BoxSet B should (coarsely) cover the manifold.

GAIO.cover_roots — Method

cover_roots(g, Dg, B::BoxSet; steps=12) -> BoxSet

Compute a covering of the roots of g within the partition P. Generally, B should be a box set containing the whole partition P, ie B = cover(P, :), and should contain a root of g.

GAIO.depth — Method

depth(tree::TreePartition)

Return the depth of the tree structure.

GAIO.expon — Function

expon(h, k=1, ϵ=0.2, δ=0.1)

Return a rough estimate of how many Newton steps should be taken, given a step size h.

GAIO.find_at_depth — Method

find_at_depth(tree, depth)

Return all node indices at a specified depth.

GAIO.finite_time_lyapunov_exponents — Method

finite_time_lyapunov_exponents(f, Df, μ::BoxFun; n=8) -> σ

Compute the Lyapunov exponents using a spatial integration method [1] based on Birkhoff's ergodic theorem. Computes

\[\sigma_j = \frac{1}{n} \int \log R_{jj}( Df^n (x) ) \, dμ (x), \quad j = 1, \ldots, d\]

with respect to an ergodic invariant measure $\mu$.

[1] Beyn, WJ., Lust, A. A hybrid method for computing Lyapunov exponents. Numer. Math. 113, 357–375 (2009). https://doi.org/10.1007/s00211-009-0236-4

GAIO.finite_time_lyapunov_exponents — Method

finite_time_lyapunov_exponents(F::SampledBoxMap, boxset::BoxSet) -> BoxFun

Compute the Finite Time Lyapunov Exponent for every box in boxset, where F represents a time-T integration of some continuous dynamical system. It is assumed that all boxes in boxset have radii of some fixed order ϵ.

GAIO.fixqr! — Method

fixqr!(Q, R)

Adjust a QR-decomposition such that the R-factor has positive diagonal entries.

GAIO.hidden_keys — Method

hidden_keys(tree)

Return all keys within the tree, including keys not corresponding to leaf nodes.

GAIO.index_pair — Method

index_pair(F::BoxMap, N::BoxSet) -> (P₁, P₀)

Compute an index pair of BoxSets P₀ ⊆ P₁ ⊆ M where M = N ∪ nbhd(N).

GAIO.index_quad — Method

index_quad(F::BoxMap, N::BoxSet) -> (P₁, P₀, P̄₁, P̄₀)

Compute a tuple of index pairs such that F: (P₁, P₀) → (P̄₁, P̄₀)

GAIO.index_to_key — Method

index_to_key(iterable, i)

Return the object held in the ith position of iterable. Used to enumerate BoxSets as $\left\{ B_1, B_2, \ldots, B_n \right\}$ in TransferOperator, BoxGraph.

GAIO.key_to_box — Method

key_to_box(P::BoxPartition, key)

Return the box associated with the index within a BoxPartition.

GAIO.key_to_index — Method

key_to_index(iterable, key)

Find the index in 1..length(iterable) which holds key, or return nothing. Used to enumerate BoxSets as $\left\{ B_1, B_2, \ldots, B_n \right\}$ in TransferOperator, BoxGraph.

GAIO.leaves — Method

leaves(tree, initial_node_idx=1)

Return the node indices of all leaves. Begins search at initial_node_idx, i.e. only returns node indices of nodes below initial_node_idx within the tree.

GAIO.linreg — Method

linreg(xs, ys)

Simple one-dimensional lunear regression used to approximate box dimension.

GAIO.marginal — Method

marginal(μ::BoxFun{Box{N}}; dim) -> BoxFun{Box{N-1}}

Compute the marginal distribution of μ along an axis given by its dimension dim.

GAIO.marginal — Method

marginal(P::BoxPartition{N}; dim) -> BoxPartition{N-1}

Construct the projection of a BoxPartition along an axis given by its dimension dim.

GAIO.marginal — Method

marginal(B::BoxSet{Box{N}}; dim) -> BoxSet{Box{N-1}}

Construct the projection of the BoxSet along an axis given by its dimension dim. This means that all boxes are projected to dimension N-1. Overlapping boxes are counted only once.

GAIO.maximal_backward_invariant_set — Function

relative_attractor(F::BoxMap, B::BoxSet; steps=12) -> BoxSet
maximal_backward_invariant_set(F::BoxMap, B::BoxSet; steps=12) -> BoxSet

Compute the attractor relative to B. B should be a (coarse) covering of the relative attractor, e.g. B = cover(P, :) for a partition P.

GAIO.morse_adjacencies — Method

Given a strong_components_output from MatrixNetworks (in particular the component map) as well as the morse map (see morse_map), compute the adjacency matrix for the morse graph.

GAIO.morse_adjacencies_and_tiles — Method

Given a transfer operator (interpreted as a transfer graph), compute the adjacency matrix for the mose graph as well as the boxes representing the vertices for the morse graph.

GAIO.morse_component_map — Method

Concatenation of the condensation map and morse map. See morse_map

GAIO.morse_map — Method

Given a Strong_components_output from MatrixNetworks (in particular the component map), compute a second map on the vertices of the condensation graph to the vertices of the morse graph. Vertices of the condensation graph which do not correspond to morse sets, get sent to the (arbitrary) vertex index 0.

 origninal         condensation        morse
 graph             graph               graph

    * ──────┐
            │
    * ──────┴───────→ * ───────────────→ *

    * ──────────────→ * ───┐     ┌─────→ *
                           │     │
    * ──────────────→ * ───┴─────┼────→  0
                                 │
    * ─────┬────────→ * ─────────┘
           │
    * ─────┤
           │
    * ─────┘

    ⋮ ==============⟹ ⋮ ==============⟹ ⋮
        condensation        morse
        map                 map

.

GAIO.morse_tiles — Method

Given a transfer operator and a morse component map (see morse_component_map), compute the boxes corresponding to the vertices of the morse graph.

GAIO.neighborhood — Method

neighborhood(B::BoxSet) -> BoxSet
nbhd(B::BoxSet) -> BoxSet

Return a one-box wide neighborhood of a BoxSet B.

GAIO.nth_iterate_jacobian — Method

nth_iterate_jacobian(f, Df, x, n; return_QR=false) -> Z[, R]

Compute the Jacobian of the n-times iterated function f ∘ f ∘ ... ∘ f at x using a QR iteration based on [1]. Requires an approximation Df of the jacobian of f, e.g. Df(x) = ForwardDiff.jacobian(f, x). Optionally, return the QR decomposition.

[1] Dieci, L., Russell, R. D., Van Vleck, E. S.: "On the Computation of Lyapunov Exponents for Continuous Dynamical Systems," submitted to SIAM J. Numer. Ana. (1993).

GAIO.point_to_box — Method

point_to_box(P::AbstractBoxPartition, point)

Find the box within a BoxPartition containing a point.

GAIO.point_to_key — Method

point_to_key(P::BoxPartition, point)

Find the index for the box within a BoxPartition contatining a point, or nothing if the point does not lie in the domain.

GAIO.preimage — Method

preimage(F::BoxMap, B::BoxSet, Q::BoxSet) -> BoxSet

Compute the (restricted to Q) preimage of B under F, i.e.

\[F^{-1} (B) \cap Q . \]

Note that the larger $Q$ is, the more calculation time required.

GAIO.preimage — Method

preimage(F::BoxMap, B::BoxSet) -> BoxSet

Efficiently compute

\[F^{-1} (B) \cap B . \]

Significantly faster than calling preimage(F, B, B).

This is not the entire preimage in the mathematical sense!

preimage(F, B) computes the RESTRICTED preimage $F^{-1} (B) \cap B$, NOT the full preimage $F^{-1} (B)$.

GAIO.radius — Method

radius(b::Box)

Return the radius of a box.

GAIO.rescale — Method

rescale(center, radius, points)

Return an iterable which calls rescale(center, radius, point) for each point in points.

GAIO.rescale — Method

rescale(points)

Return a function

(center, radius) -> rescale(center, radius, points)

Used in domain_points for BoxMap, PointDiscretizedMap.

GAIO.rescale — Method

rescale(box, point::Union{<:StaticVector{N,T}, <:NTuple{N,T}})
rescale(center, radius, point::Union{<:StaticVector{N,T}, <:NTuple{N,T}})

Scale a point within the unit box $[-1, 1]^N$ to lie within box = Box(center, radius).

GAIO.rk4 — Method

rk4(f, x, τ)

Compute one step with step size τ of the classic fourth order Runge-Kutta method.

GAIO.rk4_flow_map — Function

rk4_flow_map(f, x, step_size=0.01, steps=20)

Perform steps steps of the classic Runge-Kutta fourth order method, with step size step_size.

GAIO.sample_adaptive — Method

sample_adaptive(f, center::SVector, radius::SVector)

Create a grid of test points using the adaptive technique described in

Oliver Junge. “Rigorous discretization of subdivision techniques”. In: International Conference on Differential Equations. Ed. by B. Fiedler, K. Gröger, and J. Sprekels. 1999.

GAIO.seba — Method

seba(V::Matrix{<:Real}, Rinit=nothing, maxiter=5000) -> S, R

Construct a sparse approximation of the basis V, as described in [1]. Returns matrices $S$, $R$ such that

\[\frac{1}{2} \| V - SR \|_F^2 + \mu \| S \|_{1,1}\]

is minimized, where $\mu \in \mathbb{R}$, $\| \cdot \|_F$ is the Frobenuius-norm, and $\| \cdot \|_{1,1}$ is the element sum norm, and $R$ is orthogonal. See [1] for further information on the argument Rinit, as well as a description of the algorithm.

[1] Gary Froyland, Christopher P. Rock, and Konstantinos Sakellariou. Sparse eigenbasis approximation: multiple feature extraction across spatiotemporal scales with application to coherent set identification. Communications in Nonlinear Science and Numerical Simulation, 77:81-107, 2019. https://arxiv.org/abs/1812.02787

GAIO.seba — Method

seba(V::Vector{<:BoxFun}, Rinit=nothing; which=partition_disjoint, maxiter=5000) -> S, A

Construct a sparse eigenbasis approximation of V, as described in [1]. Returns an Array of BoxFuns corresponding to the eigenbasis, as well as a maximum-likelihood BoxFun that maps a box to the element of S which has the largest value over the support.

The keyword which is used to set the threshholding heuristic, which is used to extract a partition of the supports from the sparse basis. Builtin options are

partition_unity, partition_disjoint, partition_likelihood

which are all exported functions.

GAIO.subdivide! — Method

subdivide!(tree::TreePartition, key::keytype(tree)) -> TreePartition
subdivide!(tree::TreePartition, depth::Integer) -> TreePartition

subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, key) -> BoxSet
subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, depth) -> BoxSet

Subdivide a TreePartition at key. Dimension along which the node is subdivided depends on the depth of the node.

GAIO.subdivide — Method

subdivide(B::BoxSet{<:Any,<:Any,<:TreePartition}) -> BoxSet

Bisect every box in boxset along an axis, giving rise to a new partition of the domain, with double the amount of boxes. Axis along which to bisect depends on the depth of the nodes.

GAIO.symmetric_image — Method

symmetric_image(F::BoxMap, B::BoxSet) -> BoxSet

Efficiently compute

\[F (B) \cap B \cap F^{-1} (B) . \]

Internally performs the following computation (though more efficiently)

# create a measure with support over B
μ = BoxFun(B)

# compute transfer weights (restricted to B)
T = TransferOperator(F, B, B)

C⁺ = BoxSet(T*μ)    # support of pushforward measure
C⁻ = BoxSet(T'μ)    # support of pullback measure

C = C⁺ ∩ C⁻

GAIO.tree_search — Method

tree_search(tree, point, max_depth=Inf) -> key, node_idx

Find the key and correspoinding node index within the tree data structure containing a point.

GAIO.unstable_set — Method

unstable_set(F::BoxMap, B::BoxSet) -> BoxSet

Compute the unstable set for a box set B. Generally, B should be a small box surrounding a fixed point of F. The partition must be fine enough, since no subdivision occurs in this algorithm.

GAIO.vertices — Method

vertices(box)

Return an iterator over the vertices of a box = Box(center, radius).

GAIO.@save — Macro

@save boxset prefix="./" suffix=".boxset" -> filename
@save boxset filename -> filename

Save a BoxSet as a list of keys. The default file name is the variable name.

Note that this does not include information on the partition of the BoxSet, just the keys.

.

@save boxmap source prefix="./" suffix=".boxmap" -> filename
@save boxmap source filename -> filename

Save a BoxMap as a list of source-keys and their image-keys in the form

key_1 -> {image_1, image_2, image_3}
key_2 -> {image_2, image_4, image_8, image_6}
⋮

.

@save transfer_operator prefix="./" suffix=".boxmap" -> filename
@save transfer_operator filename -> filename

Save a TransferOperator as a list of keys and their image-keys in the form

key_1 -> {image_1, image_2, image_3}
key_2 -> {image_2, image_4, image_8, image_6}
⋮