Plotting

GAIO.jl offers plotting recipes for both Plots.jl and Makie.jl. This means that one can use all the surrounding functionality of Plots.jl or makie.jl, e.g. creating multiple subplots, animations, colorbars, etc. simply by loading either Plots.jl or Makie.jl.

Why offer both Plots and Makie recipes?

Makie.jl offers much more funcitonality and performance in interactive plots, but suffers much more greatly from the well known time to fist plot problem. Hence GAIO.jl offers Plots.jl recipes for fast, 2-dimensional, non-interactive plotting, and Makie.jl recipes for interactive, 2- and 3-dimensional, or publication-quality visualizations. To see a difference in the plotting capability, see the Hénon map example.

Using Plots

To make a plot, one needs to simply load the package. Add Plots using the package manager

pkg> add Plots

Load Plots with

using Plots

Using Makie

To see a plot, one needs to load one of the Makie backends (see makie documentation about backends). We will use GLMakie, which uses OpenGL. Add GLMakie using the package manager

pkg> add GLMakie

Load the GLMakie backend and some plotting tools with

using GLMakie: plot #, Axis3, Colorbar, etc...

A note on Namespaces

Makie and GAIO.jl both export the type Box. For this reason, it is recommended NOT to use

using GLMakie

and instead only load the function names one needs from Makie.

Common Interface

To plot a BoxSet or BoxFun b, simply call

plot(b)

The mutating function plot! is also available. All of Plots.jl's or Makie.jl's keyword arguments, such as color, colormap, etc. can be used. In addition, the keyword argument projection is used to project to a lower dimensional space if the dimension of the space $d$ is greater than 2 for Plots.jl or greater than 3 for Makie.jl. By default, the function used is x -> x[1:2] for Plots.jl and x -> x[1:3] for Makie.jl. For an example using a custom projection function, eg. to plot the unstable set of the dadras system:

using GAIO
using GLMakie: plot

# the Dadras system
const a, b, c = 8.0, 40.0, 14.9
v((x,y,z,w)) = (a*x-y*z+w, x*z-b*y, x*y-c*z+x*w, -y)
f(x) = rk4_flow_map(v, x, 0.01, 5)

cen, rad = (0,0,0,0), (250,150,200,25)
P = BoxPartition(Box(cen, rad), (128,128,128,128))
F = AdaptiveBoxMap(f, P.domain)

x = zeros(4)        # equilibrium
W = unstable_set(F, P[x])

A = [0 1 0 0;
     0 0 1 0;
     0 0 0 1]

plot(W, projection = x -> A*x)

GAIO.plotboxes — Function

plot(boxset::BoxSet)
plot(boxfun::BoxFun)
plot!(boxset::BoxSet)
plot!(boxfun::BoxFun)

Plot a BoxSet or BoxFun.

Special Attributes:

projection = x -> x[1:3] If the dimension of the system is larger than 3, use this function to project to 3-d space.

color = :red Color used for the boxes.

colormap = :default Colormap used for plotting BoxFuns values.

marker = HyperRectangle(GeometryBasics.Vec3f0(0), GeometryBasics.Vec3f0(1)) The marker for an individual box. Only works if using Makie for plotting.

All other attributes are taken from MeshScatter.