Data Structures

GAIO.Box

GAIO.BoxGraph

GAIO.BoxPartition

GAIO.BoxSet

GAIO.TreePartition

GAIO.bounded_point_to_key

GAIO.key_to_box

GAIO.point_to_box

GAIO.point_to_key

GAIO.subdivide

GAIO.subdivide!

GAIO.volume

GAIO.Box — Type

Box{N,T}(center, radius)
Box(center, radius)

A generalized box in dimension N with element type T. Mathematically, this is a set

\[[center_1 - radius_1,\ center_1 + radius_1) \ \times \ \ldots \ \times \ [center_N - radius_N,\ center_N + radius_N)\]

Fields:

center: vector where the box's center is located

radius: vector of radii, length of the box in each dimension

Methods implemented:

:(==), in #, etc ...

.

GAIO.volume — Method

volume(box::Box)

Compute the volume of a box.

GAIO.BoxPartition — Type

BoxPartition(domain::Box{N}, dims::NTuple{N,<:Integer} = ntuple(_->1, N))

Data structure to partition a domain into a dims[1] x dims[2] x ... dims[N] equidistant box grid.

Fields:

domain: box defining the entire domain

left: leftmost / bottom edge of the domain

scale: 1 / diameter of each box in the new partition (componentwise)

dims: tuple, number of boxes in each dimension

Methods implemented:

:(==), ndims, size, length, keys, keytype #, etc ...

.

GAIO.point_to_key — Function

point_to_key(P::BoxPartition, point)

Find the index for the box within a BoxPartition contatining a point, or nothing if the point does not lie in the domain.

GAIO.bounded_point_to_key — Function

bounded_point_to_key(P::BoxPartition, point)

Find the cartesian index of the nearest box within a BoxPartition to a point. Conicides with point_to_key if the point lies in the partition. Default behavior is to set NaN = Inf if NaNs are present in point.

GAIO.key_to_box — Function

key_to_box(P::BoxPartition, key)

Return the box associated with the index within a BoxPartition.

GAIO.point_to_box — Function

point_to_box(P::AbstractBoxPartition, point)

Find the box within a BoxPartition containing a point.

GAIO.subdivide — Method

subdivide(P::BoxPartition, dim) -> BoxPartition
subdivide(B::BoxSet, dim) -> BoxSet

Bisect every box in the BoxPartition or BoxSet along the axis dim, giving rise to a new partition of the domain, with double the amount of boxes.

GAIO.TreePartition — Type

TreePartition(domain::Box)

Binary tree structure to partition domain into (variably sized) boxes.

Fields:

domain: Box denoting the full domain.

nodes: vector of Nodes. Each node holds two indices pointing to other nodes in the vector, or 0 if the node is a leaf.

Methods implemented:

copy, keytype, keys, subdivide #, etc...

.

GAIO.subdivide! — Function

subdivide!(tree::TreePartition, key::keytype(tree)) -> TreePartition
subdivide!(tree::TreePartition, depth::Integer) -> TreePartition

subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, key) -> BoxSet
subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, depth) -> BoxSet

Subdivide a TreePartition at key. Dimension along which the node is subdivided depends on the depth of the node.

GAIO.BoxSet — Type

BoxSet(partition, indices::AbstractSet)

Internal data structure to hold boxes within a partition.

Constructors:

set of all boxes in partition / box set P:

B = cover(P, :)    

cover the point x, or points x = [x_1, x_2, x_3] # etc ... using boxes from P

B = cover(P, x)

a covering of S using boxes from P

S = [Box(center_1, radius_1), Box(center_2, radius_2), Box(center_3, radius_3)] # etc... 
B = cover(P, S)

Fields:

partition: the partition that the set is defined over

set: set of partition-keys corresponding to the boxes in the set

Most set operations such as

union, intersect, setdiff, symdiff, issubset, isdisjoint, issetequal, isempty, length, # etc...

are supported.

GAIO.BoxGraph — Type

Graph(gstar::TransferOperator) -> BoxGraph
Graph(g::BoxMap, boxset::BoxSet) = Graph(TransferOperator(g, boxset, boxset))

Directed Graph representation of a TransferOperator. The boxes in a BoxSet are enumerated as in TransferOperator. This means if the domain, codomain are taken from a TranferOperator, then the graph vertices are numbered 1 .. length(domain ∪ codomain).

A directed edge exists from the i'th box b_i to the j'th box b_j if the BoxMap g has b_j ∩ g⁻¹(b_i) ≠ ∅. Equivalently,

has_edge(g::BoxGraph, i, j) = !iszero( Matrix(g.gstar)[j,i] )

Graphs.jl operations like

vertices, edges, weights, inneighbors, outneighbors, # etc...

are supported. Algorithms in Graphs.jl should work "out of the box", but will return whatever Graphs.jl returns by default. To convert a (integer) vertex index from the graph into a box index from the partition, one can call

BoxSet(boxgraph, graph_index_or_indices)

If you would like to see specific behavior implemented, please open an issue!

Implementation details, not important for use:

We want to turn a matrix representation

        domain -->
codomain  .   .   .   .   .
    |     .   .   .   .   .
    |     .   .   .   .   .
    v     .   .  mat  .   .
          .   .   .   .   .
          .   .   .   .   .

into a graph representation

  domain ∪ codomain
  .---------.   .
 / \       /   /
.   .-----.---.

!! efficiently !!

Julia's Graphs package only allows integer-indexed vertices so we need to enumerate domain ∪ codomain. To do this, we enumerate the domain, then skip the boxes in the codomain which are already in the domain, then continue enumerating the rest of the codomain.

We therefore permute the row indices of the weight matrix so that the skipped elements of the codomain come first.