Data Structures

GAIO.Box

GAIO.BoxFun

GAIO.BoxGraph

GAIO.BoxPartition

GAIO.BoxSet

GAIO.TransferOperator

GAIO.TreePartition

Base.:∘

Base.sum

GAIO.bounded_point_to_key

GAIO.key_to_box

GAIO.point_to_box

GAIO.point_to_key

GAIO.subdivide

GAIO.subdivide!

GAIO.volume

GAIO.Box — Type

Box{N,T}(center, radius)
Box(center, radius)

A generalized box in dimension N with element type T. Mathematically, this is a set

\[[center_1 - radius_1,\ center_1 + radius_1) \ \times \ [center_N - radius_N,\ center_N + radius_N)\]

Fields:

center: vector where the box's center is located

radius: vector of radii, length of the box in each dimension

Methods implemented:

:(==), in #, etc ...

.

GAIO.volume — Method

volume(box::Box)

Compute the volume of a box.

GAIO.BoxPartition — Type

BoxPartition(domain::Box{N}, dims::NTuple{N,<:Integer} = ntuple(_->1, N))

Data structure to partition a domain into a dims[1] x dims[2] x ... dims[N] equidistant box grid.

Fields:

domain: box defining the entire domain

left: leftmost / bottom edge of the domain

scale: 1 / diameter of each box in the new partition (componentwise)

dims: tuple, number of boxes in each dimension

Methods implemented:

:(==), ndims, size, length, keys, keytype #, etc ...

.

GAIO.point_to_key — Function

point_to_key(P::BoxPartition, point)

Find the index for the box within a BoxPartition contatining a point, or nothing if the point does not lie in the domain.

GAIO.bounded_point_to_key — Function

bounded_point_to_key(partition::BoxPartition, point)

Find the cartesian index of the nearest box within a BoxPartition to a point. Conicides with point_to_key if the point lies in the partition. Default behavior is to set NaN = Inf if NaNs are present in point.

GAIO.key_to_box — Function

key_to_box(P::BoxPartition, key)

Return the box associated with the index within a BoxPartition.

GAIO.point_to_box — Function

point_to_box(P::AbstractBoxPartition, point)

Find the box within a BoxPartition containing a point.

GAIO.subdivide — Method

subdivide(P::BoxPartition, dim) -> BoxPartition
subdivide(B::BoxSet, dim) -> BoxSet

Bisect every box in the BoxPartition or BoxSet along the axis dim, giving rise to a new partition of the domain, with double the amount of boxes.

GAIO.TreePartition — Type

TreePartition(domain::Box)

Binary tree structure to partition domain into (variably sized) boxes.

Fields:

domain: Box denoting the full domain.

nodes: vector of Nodes. Each node holds two indices pointing to other nodes in the vector, or 0 if the node is a leaf.

Methods implemented:

copy, keytype, keys, subdivide #, etc...

.

GAIO.subdivide! — Function

subdivide!(tree::TreePartition, key::keytype(tree)) -> TreePartition
subdivide!(tree::TreePartition, depth::Integer) -> TreePartition

subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, key) -> BoxSet
subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, depth) -> BoxSet

Subdivide a TreePartition at key. Dimension along which the node is subdivided depends on the depth of the node.

Missing docstring.

Missing docstring for BoxMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MonteCarloBoxMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GridBoxMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PointDiscretizedBoxMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AdaptiveBoxMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SampledBoxMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IntervalBoxMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CPUSampledBoxMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUSampledBoxMap. Check Documenter's build log for details.

GAIO.BoxSet — Type

BoxSet(partition, indices::AbstractSet)

Internal data structure to hold boxes within a partition.

Constructors (all constructors work with box sets as well):

For all boxes in a partition:

partition[:]

For one box containing a point x:

partition[x]

For a covering of an iterable S = [Box(c_1, r_1), Box(c_2, r_r)] # etc...:

partition[S]

Fields:

partition: the partition that the set is defined over

set: set of partition-keys corresponding to the boxes in the set

Most set operations such as

union, intersect, setdiff, symdiff, issubset, isdisjoint, issetequal, isempty, length, # etc...

are supported.

GAIO.TransferOperator — Type

TransferOperator(map::BoxMap, domain::BoxSet)
TransferOperator(map::BoxMap, domain::BoxSet, codomain::BoxSet)

Discretization of the Perron-Frobenius operator, or transfer operator. Implemented as a sparse matrix with indices referring to two BoxSets: domain and codomain.

There exists two constructors:

only provide a boxmap and a domain. In this case, the codomain is generated as the image of domain under the boxmap.
julia> P = BoxPartition( Box((0,0), (1,0)), (10,10) )
  10 x 10 - element BoxPartition

julia> domain = BoxSet( P, Set((1,2), (2,3), (3,4)) )
  3 - element Boxset over 10 x 10 - element BoxPartition

julia> T = TransferOperator(boxmap, domain)
  TransferOperator over [...]

provide domain and codomain. In this case, the size of the transition matrix is given.
julia> codomain = domain
  3 - element Boxset over 10 x 10 - element BoxPartition

julia> T = TransferOperator(boxmap, domain, codomain)
  TransferOperator over [...]

Fields:

mat: SparseMatrixCSC containing transfer weights. The index T.mat[i,j] represents the transfer weight FROM the j'th box in codomain TO the i'th box in domain.

boxmap: SampledBoxMap map which dictates the transfer weights.

domain: BoxSet which contains keys for the already calculated transfers. Effectively, these are column pointers, i.e. the jth column of T.mat contains transfer weights FROM box Bj, where Bj is the jth box of domain.

codomain: BoxSet which contains keys for the already calculated transfers. Effectively, these are row pointers, i.e. the ith row of T.mat contains transfer weights TO box Bi, where Bi is the ith box of codomain.

        domain -->
codomain  .   .   .   .   .
    |     .   .   .   .   .
    |     .   .   .   .   .
    v     .   .  mat  .   .
          .   .   .   .   .
          .   .   .   .   .
          .   .   .   .   .
          .   .   .   .   .

It is important to note that TranferOperator is only supported over the box set domain, but if one lets a TranferOperator act on a BoxFun, e.g. by multiplication, then the domain is extended "on the fly" to include the support of the BoxFun.

Methods Implemented:

:(==), axes, size, eltype, getindex, setindex!, SparseArrays.sparse, Arpack.eigs, LinearAlgebra.mul! #, etc ...

Implementation detail:

The reader may have noticed that the matrix representation depends on the order of boxes in support. For this reason an OrderedSet is used. BoxSets using regular Sets will be copied and converted to OrderedSets.

.

GAIO.BoxFun — Type

BoxFun(partition, vals)

Discretization of a measure over the domain partition.domain, as a piecewise constant function over the boxes of partition.

Implemented as a sparse vector over the indices of partition.

Fields:

partition: An AbstractBoxPartition whose indices are used

for vals

vals: A dictionary whose keys are the box indices from

partition, and whose values represent the values of the function.

Methods implemented:

length, sum, iterate, values, isapprox, ∘, LinearAlgebra.norm, LinearAlgebra.normalize!

.

Base.sum — Method

sum(f, μ::BoxFun)
sum(f, μ::BoxFun, B::BoxSet)
μ(B) = sum(identity, μ, B)

Integrate a function f using μ as a density, that is, if boxfun is the discretization of a measure $\mu$ over the domain $Q$, then approximate the value of

\[\int_Q f \, d\mu .\]

If a BoxSet B is passed as the third argument, then the integration is restricted to the boxes in B

\[\int_{Q \cap \bigcup_{b \in B} b} f \, d\mu .\]

The notation μ(B) is offered to compute $\mu (\bigcup_{b \in B} b)$.

Base.:∘ — Method

∘(f, boxfun::BoxFun) -> BoxFun
∘(boxfun::BoxFun, F::BoxMap) -> BoxFun

Postcompose the function f with the boxfun, or precompose a BoxMap F with the boxfun (by applying the Koopman operator). Note that the support of BoxFun must be forward-invariant under F.

GAIO.BoxGraph — Type

Graph(gstar::TransferOperator) -> BoxGraph
Graph(g::BoxMap, boxset::BoxSet) = Graph(TransferOperator(g, boxset, boxset))

Directed Graph representation of a TransferOperator. The boxes in a BoxSet are enumerated as in TransferOperator. This means if the domain, codomain are taken from a TranferOperator, then the graph vertices are numbered 1 .. length(domain ∪ codomain).

A directed edge exists from the i'th box b_i to the j'th box b_j if the BoxMap g has b_j ∩ g⁻¹(b_i) ≠ ∅. Equivalently,

has_edge(g::BoxGraph, i, j) = !iszero( Matrix(g.gstar)[j,i] )

Graphs.jl operations like

vertices, edges, weights, inneighbors, outneighbors, # etc...

are supported. Algorithms in Graphs.jl should work "out of the box", but will return whatever Graphs.jl returns by default. To convert a (integer) vertex index from the graph into a box index from the partition, one can call

BoxSet(boxgraph, graph_index_or_indices)

If you would like to see specific behavior implemented, please open an issue!

Implementation details, not important for use:

We want to turn a matrix representation

        domain -->
codomain  .   .   .   .   .
    |     .   .   .   .   .
    |     .   .   .   .   .
    v     .   .  mat  .   .
          .   .   .   .   .
          .   .   .   .   .

into a graph representation

  domain ∪ codomain
  .---------.   .
 / \       /   /
.   .-----.---.

!! efficiently !!

Julia's Graphs package only allows integer-indexed vertices so we need to enumerate domain ∪ codomain. To do this, we enumerate the domain, then skip the boxes in the codomain which are already in the domain, then continue enumerating the rest of the codomain.

We therefore permute the row indices of the weight matrix so that the skipped elements of the codomain come first.