Data Structures

GAIO.Box

GAIO.BoxFun

GAIO.BoxGraph

GAIO.BoxMap

GAIO.BoxPartition

GAIO.BoxSet

GAIO.SampledBoxMap

GAIO.TransferOperator

GAIO.TreePartition

Base.:∘

Base.sum

GAIO.AdaptiveBoxMap

GAIO.PointDiscretizedMap

GAIO.subdivide

GAIO.subdivide!

GAIO.volume

GAIO.Box — Type

Box{N,T}(center, radius)
Box(center, radius)

A generalized box in dimension N with element type T. Mathematically, this is a set

\[[center_1 - radius_1,\ center_1 + radius_1) \ \times \ [center_N - radius_N,\ center_N + radius_N)\]

Fields:

center: vector where the box's center is located

radius: vector of radii, length of the box in each dimension

Methods implemented:

:(==), in #, etc ...

.

GAIO.volume — Method

Computes the volume of a box.

GAIO.BoxPartition — Type

BoxPartition(
    domain::Box{N}, 
    dims::NTuple{N,<:Integer} = ntuple(_->1, N)
    ; indextype = (N < 9 : IndexLinear() : IndexCartesian())
)

Data structure to partition a domain into a dims[1] x dims[2] x ... dims[N] equidistant box grid.

Fields:

domain: box defining the entire domain

left: leftmost / bottom edge of the domain

scale: 1 / diameter of each box in the new partition (componentwise)

dims: tuple, number of boxes in each dimension

dimsprod: for indexing the partition. BoxPartition uses linear indices, i.e. keys are counted up in first dimension first, then second dimension, etc...

indextype: whether the partition will use cartesian or linear indices by default. Values can be IndexLinear(), IndexCartesian()

Methods implemented:

:(==), ndims, size, length, keys, keytype #, etc ...

.

GAIO.subdivide — Method

subdivide(P::BoxPartition, dim) -> BoxPartition
subdivide(B::BoxSet, dim) -> BoxSet

Bisect every box in boxset along the axis dim, giving rise to a new partition of the domain, with double the amount of boxes.

GAIO.TreePartition — Type

TreePartition(domain::Box)

Binary tree structure to partition domain into (variably sized) boxes.

Fields:

domain: Box denoting the full domain.

nodes: vector of Nodes. Each node holds two indices pointing to

other nodes in the vector, or 0.

regular_partitions: vector of BoxPartitions. The indices held in

a node also refer to this vector.

Methods implemented:

copy, keytype #, etc...

.

GAIO.subdivide! — Function

subdivide!(tree:TreePartition), key::NTuple{2,<:Integer}) -> TreePartition
subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, key::NTuple{2,<:Integer}) -> BoxSet

Subdivide a TreePartition at the node key. Dimension along which the node is subdivided depends on the depth of the node.

GAIO.SampledBoxMap — Type

SampledBoxMap(map, domain::Box, domain_points, image_points, acceleration)

Transforms a $map: Q → Q$ defined on points in the box $Q ⊂ ℝᴺ$ to a SampledBoxMap defined on Boxes.

Constructors:

BoxMap

PointDiscretizedMap

AdaptiveBoxMap

Fields:

map: map that defines the dynamical system.

domain: domain of the map, B.

domain_points: the spread of test points to be mapped forward in intersection algorithms. Must have the signature domain_points(center, radius) and return an iterator of points within Box(center, radius).

image_points: the spread of test points for comparison in intersection algorithms. Must have the signature domain_points(center, radius) and return an iterator of points within Box(center, radius).

acceleration: Whether to use optimized functions in intersection algorithms. Accepted values: nothing, BoxMapCPUCache, BoxMapGPUCache. BoxMapGPUCache does nothing unless you have a CUDA capable gpu.

.

GAIO.BoxMap — Type

BoxMap(map, domain::Box{N,T}, accel=nothing; no_of_points=4*N*pick_vector_width(T)) -> SampledBoxMap
BoxMap(map, P::BoxPartition{N,T}, accel=nothing; no_of_points=4*N*pick_vector_width(T)) -> SampledBoxMap

Construct a SampledBoxMap which uses no_of_points Monte-Carlo test points.

GAIO.PointDiscretizedMap — Function

PointDiscretizedMap(map, domain, points, accel=nothing) -> SampledBoxMap

Construct a SampledBoxMap that uses the iterator points as test points. points must be an array or iterator of test points within the unit cube $[-1,1]^N$.

GAIO.AdaptiveBoxMap — Function

AdaptiveBoxMap(f, domain::Box, accel=nothing) -> SampledBoxMap

Construct a SampledBoxMap which uses sample_adaptive to generate test points.

GAIO.BoxSet — Type

BoxSet(partition, indices::AbstractSet)

Internal data structure to hold boxes within a partition.

Constructors:

For all boxes in a partition:

partition[:]

For one box containing a point x:

partition[x]

For a covering of an iterable S = [Box(c_1, r_1), Box(c_2, r_r)] # etc...:

partition[S]

Fields:

partition: the partition that the set is defined over

set: set of partition-keys corresponding to the boxes in the set

Most set operations such as

union, intersect, setdiff, symdiff, issubset, isdisjoint, issetequal, isempty, length, # etc...

are supported.

GAIO.TransferOperator — Type

TransferOperator(map::BoxMap, support::BoxSet)

Discretization of the Perron-Frobenius operator, or transfer operator. Implemented as a sparse matrix with the same linear indices as support, e.g. if

julia> B = BoxSet(partition, [3,10,30])
  Boxset over [...] partition

julia> T = TransferOperator(boxmap, B)
  TransferOperator over [...] BoxSet

for some partition and boxmap, then

julia> axes(T)
  ([3, 10, 30], [3, 10, 30])

Fields:

boxmap: SampledBoxMap map which relates to the transfers.

support: BoxSet which contains keys for the already calculated transfers. Effectively, these are row/column pointers, i.e. the first column of T.mat contains transfer weights FROM box B1, where B1 is the first box of support.

variant_set: BoxSet which contains keys for potential boxes lying outside of support, i.e. it could be that support is not an invariant set. In this case, variant_set contains the boxes which were not in support, but whose preimage lies in support.

mat: SparseMatrixCSC containing transfer weights. The index T.mat[i,j] represents the transfer weight FROM the j'th box in support TO the i'th box in support. If support is not invariant, then the matrix will be tall. In this case the rows are counted in support and then in variant_set.

support -->
  |     .   .   .   .   .
  |     .   .   .   .   .
  v     .   .   .   .   .
        .   .  mat  .   .
        .   .   .   .   .
 var-   .   .   .   .   .
 ian-   .   .   .   .   .
 t_set  .   .   .   .   .     
  |     .   .   .   .   .
  |     .   .   .   .   .
  v     .   .   .   .   .

It is important to note that TranferOperator is only supported over the box set B, but if one lets a TranferOperator act on a BoxFun, then the support B is extended "on the fly" to include the support of the BoxFun.

Methods Implemented:

:(==), axes, size, eltype, getindex, setindex!, SparseArrays.sparse, Arpack.eigs, LinearAlgebra.mul! #, etc ...

Implementation detail:

The reader may have noticed that the matrix representation depends on the order of boxes in support. For this reason an OrderedSet is used. BoxSets using regular Sets will be copied and converted to OrderedSets.

.

GAIO.BoxFun — Type

BoxFun(partition, vals)

Discretization of a measure over the domain partition.domain, as a piecewise constant function over the boxes of partition.

Implemented as a sparse vector over the indices of partition.

Fields:

partition: An AbstractBoxPartition whose indices are used

for vals

vals: A dictionary whose keys are the box indices from

partition, and whose values represent the values of the function.

Methods implemented:

length, sum, iterate, values, isapprox, ∘, LinearAlgebra.norm, LinearAlgebra.normalize!

.

Base.sum — Method

sum(f, μ::BoxFun)
sum(f, μ::BoxFun, B::BoxSet)
μ(B) = sum(identity, μ, B)

Integrate a function f using μ as a density, that is, if boxfun is the discretization of a measure $\mu$ over the domain $Q$, then approximate the value of

\[\int_Q f \, d\mu .\]

If a BoxSet B is passed as the third argument, then the integration is restricted to the boxes in B

\[\int_{Q \cap \bigcup_{b \in B} b} f \, d\mu .\]

The notation μ(B) is offered to compute $\mu (\bigcup_{b \in B} b)$.

Base.:∘ — Method

∘(f, boxfun::BoxFun) -> BoxFun
∘(boxfun::BoxFun, F::BoxMap) -> BoxFun

Postcompose the function f with the boxfun, or precompose a BoxMap F with the boxfun (by applying the Koopman operator).

GAIO.BoxGraph — Type

Graph(gstar::TransferOperator) -> BoxGraph
Graph(g::BoxMap, boxset::BoxSet) = Graph(TransferOperator(g, boxset))

Directed Graph representation of a TransferOperator. The boxes in a BoxSet are enumerated as in TransferOperator. This means the vertices are numbered 1 .. size(gstar)[1]. A directed edge exists from the i'th box b_i to the j'th box b_j in boxgraph.gstar.support if the BoxMap g has b_j ∩ g⁻¹(b_i) ≠ ∅. Equivalently,

has_edge(g::BoxGraph, i, j) = !iszero( Matrix(g.gstar)[j,i] )

Graphs.jl operations like

vertices, edges, weights, inneighbors, outneighbors, # etc...

are supported. Some algorithms are adapted like

strongly_connected_components

to return custom objects. All other algorightms in Graphs.jl should work "out of the box", but will return whatever Graphs.jl returns by default. To convert a (integer) vertex index from the graph into a box index from the partition, one can call

BoxSet(boxgraph, graph_index_or_indices)

If you would like to see specific behavior implemented, please open an issue!

.