Data Structures

GAIO.Box

GAIO.BoxFun

GAIO.BoxMap

GAIO.BoxPartition

GAIO.BoxSet

GAIO.SampledBoxMap

GAIO.TransferOperator

GAIO.TreePartition

Base.:∘

Base.sum

GAIO.AdaptiveBoxMap

GAIO.PointDiscretizedMap

GAIO.subdivide

GAIO.subdivide!

GAIO.volume

GAIO.Box — Type

Box{N,T}(center, radius)
Box(center, radius)

A generalized box in dimension N with element type T. Mathematically, this is a set

\[[center_1 - radius_1,\ center_1 + radius_1) \ \times \ [center_N - radius_N,\ center_N + radius_N)\]

Fields:

center: vector where the box's center is located

radius: vector of radii, length of the box in each dimension

Methods implemented:

:(==), in #, etc ...

.

GAIO.volume — Method

Computes the volume of a box.

GAIO.BoxPartition — Type

BoxPartition(domain::Box{N}, dims::NTuple{N,<:Integer} = ntuple(_->1, N))

Data structure to partition a box into a dims[1] x dims[2] x ... dims[N] equidistant grid.

Fields:

domain: box defining the entire domain

left: leftmost / bottom edge of the domain

scale: 1 / diameter of each box in the new partition (componentwise)

dims: tuple, number of boxes in each dimension

dimsprod: for indexing the partition. BoxPartition uses linear indices, i.e. keys are counted up in first dimension first, then second dimension, etc...

         1st dim →
          * — * — * — *
    2nd   | 1 | 2 | 3 |
    dim   * — * — * — *
    ↓     | 4 | 5 | 6 |
          * — * — * — *

Methods implemented:

:(==), ndims, size, length, keys, keytype #, etc ...

.

GAIO.subdivide — Method

subdivide(P::BoxPartition, dim) -> BoxPartition
subdivide(B::BoxSet, dim) -> BoxSet

Bisect every box in boxset along the axis dim, giving rise to a new partition of the domain, with double the amount of boxes.

GAIO.TreePartition — Type

TreePartition(domain::Box)

Binary tree structure to partition domain into (variably sized) boxes.

Fields:

domain: Box denoting the full domain.

nodes: vector of Nodes. Each node holds two indices pointing to

other nodes in the vector, or 0.

regular_partitions: vector of BoxPartitions. The indices held in

a node also refer to this vector.

Methods implemented:

copy, keytype #, etc...

.

GAIO.subdivide! — Function

subdivide!(tree:TreePartition), key::NTuple{2,<:Integer}) -> TreePartition
subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, key::NTuple{2,<:Integer}) -> BoxSet

Subdivide a TreePartition at the node key. Dimension along which the node is subdivided depends on the depth of the node.

GAIO.SampledBoxMap — Type

SampledBoxMap(map, domain::Box, domain_points, image_points, acceleration)

Transforms a $map: Q → Q$ defined on points in the box $Q ⊂ ℝᴺ$ to a SampledBoxMap defined on Boxes.

Constructors:

BoxMap

PointDiscretizedMap

AdaptiveBoxMap

Fields:

map: map that defines the dynamical system.

domain: domain of the map, B.

domain_points: the spread of test points to be mapped forward in intersection algorithms. Must have the signature domain_points(center, radius) and return an iterator of points within Box(center, radius).

image_points: the spread of test points for comparison in intersection algorithms. Must have the signature domain_points(center, radius) and return an iterator of points within Box(center, radius).

acceleration: Whether to use optimized functions in intersection algorithms. Accepted values: nothing, BoxMapCPUCache, BoxMapGPUCache. BoxMapGPUCache does nothing unless you have a CUDA capable gpu.

.

GAIO.BoxMap — Type

BoxMap(map, domain::Box{N,T}, accel=nothing; no_of_points=4*N*pick_vector_width(T)) -> SampledBoxMap
BoxMap(map, P::BoxPartition{N,T}, accel=nothing; no_of_points=4*N*pick_vector_width(T)) -> SampledBoxMap

Construct a SampledBoxMap which uses no_of_points Monte-Carlo test points.

GAIO.PointDiscretizedMap — Function

PointDiscretizedMap(map, domain, points, accel=nothing) -> SampledBoxMap

Construct a SampledBoxMap that uses the iterator points as test points. points must be an array or iterator of test points within the unit cube $[-1,1]^N$.

GAIO.AdaptiveBoxMap — Function

AdaptiveBoxMap(f, domain::Box, accel=nothing) -> SampledBoxMap

Construct a SampledBoxMap which uses sample_adaptive to generate test points.

GAIO.BoxSet — Type

BoxSet(partition, indices::AbstractSet)

Internal data structure to hold boxes within a partition.

Constructors:

For all boxes in a partition:

partition[:]

For one box containing a point x:

partition[x]

For a covering of a set S = [Box(c_1, r_1), Box(c_2, r_r)] # etc...:

partition[S]

Fields:

partition: the partition that the set is defined over

set: set of partition-keys corresponding to the boxes in the set

Most set operations such as

union, intersect, setdiff, symdiff, issubset, isdisjoint, issetequal, isempty, length, # etc...

are supported.

GAIO.BoxFun — Type

BoxFun(partition, vals)

Discretization of a function over the domain partition.domain, as a piecewise constant function over the boxes of partition.

Implemented as a sparse vector over the indices of partition.

Fields:

partition: An AbstractBoxPartition whose indices are used

for vals

vals: A sparse vector whose indices are the box indices from

partition, and whose values represent the values of the function.

Methods implemented:

length, LinearAlgebra.norm, LinearAlgebra.normalize!

.

Base.sum — Method

sum(f, boxfun::BoxFun)

Integrate a function f using boxfun as a density, that is, if boxfun is the discretization of a measure $\mu$ over the domain $Q$, then approximate the value of

\[\int_Q f \, d\mu .\]

Base.:∘ — Method

∘(f, boxfun::BoxFun) -> BoxFun

Compose the function f with the boxfun.

GAIO.TransferOperator — Type

TransferOperator(map::BoxMap, support::BoxSet, mat::SparseMatrixCSC)
TransferOperator(map::BoxMap, support::BoxSet)

Discretization of the Perron-Frobenius operator, or transfer operator. Implemented as a sparse matrix with the same linear indices as support, e.g. if

julia> B = BoxSet(partition, [3,10,30])
  Boxset over [...] partition

julia> T = TransferOperator(boxmap, B)
  TransferOperator over [...] BoxSet

for some partition and boxmap, then

julia> axes(T)
  ([3, 10, 30], [3, 10, 30])

It is important to note that TranferOperator is only supported over the box set B, but if one lets a TranferOperator act on a BoxFun, then the support B is extended "on the fly" to include the support of the BoxFun.

Methods Implemented:

:(==), axes, size, eltype, getindex, setindex!, SparseArrays.sparse, Arpack.eigs, LinearAlgebra.mul! #, etc ...

.

Add SparseArrays.sparse(g::TransferOperator) once next PR is pushed