Library Reference

GAIO.BoxMap

GAIO.BoxSet

GAIO.Node

Arpack.eigs

Arpack.eigs

Arpack.eigs

Base.sum

GAIO.AdaptiveBoxMap

GAIO.GridBoxMap

GAIO.MonteCarloBoxMap

GAIO.PointDiscretizedBoxMap

GAIO.PointDiscretizedBoxMap

GAIO.PointDiscretizedBoxMap

GAIO.adaptive_newton_step

GAIO.approx_lipschitz

GAIO.bounded_point_to_key

GAIO.center

GAIO.chain_recurrent_set

GAIO.cover

GAIO.cover_boxes

GAIO.cover_roots

GAIO.depth

GAIO.find_at_depth

GAIO.finite_time_lyapunov_exponents

GAIO.hidden_keys

GAIO.index_to_key

GAIO.key_to_box

GAIO.key_to_index

GAIO.leaves

GAIO.plotboxes

GAIO.plotboxes!

GAIO.point_to_box

GAIO.point_to_key

GAIO.relative_attractor

GAIO.rescale

GAIO.rescale

GAIO.rescale

GAIO.rk4

GAIO.rk4_flow_map

GAIO.sample_adaptive

GAIO.subdivide

GAIO.subdivide!

GAIO.unstable_set

GAIO.vertices

GAIO.BoxMap — Method

BoxMap(map, domain; no_of_points=ntuple(_->4, N)) -> SampledBoxMap

Transforms a $map: Q → Q$ defined on points in the domain $Q ⊂ ℝᴺ$ to a SampledBoxMap defined on Boxes.

By default uses a grid of sample points.

GAIO.BoxSet — Method

BoxSet(boxgraph, graph_index_or_indices) -> BoxSet

Construct a BoxSet from some index / indices of vertices in a BoxGraph.

GAIO.Node — Type

Node structure used for TreePartitions

Fields:

left and right refer to indices w.r.t.

trp.nodes for a TreePartition trp.

Arpack.eigs — Function

eigs(gstar::TransferOperator [; kwargs...]) -> (d[, v], nconv)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Arpack.eigs — Function

eigs(gstar::TransferOperator [; kwargs...]) -> (d[, v], nconv)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Arpack.eigs — Function

eigs(gstar::TransferOperator [; kwargs...]) -> (d[, v], nconv)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Base.sum — Method

sum(f, μ::BoxFun)
sum(f, μ::BoxFun, B::BoxSet)
μ(B) = sum(identity, μ, B)

Integrate a function f using μ as a density, that is, if boxfun is the discretization of a measure $\mu$ over the domain $Q$, then approximate the value of

\[\int_Q f \, d\mu .\]

If a BoxSet B is passed as the third argument, then the integration is restricted to the boxes in B

\[\int_{Q \cap \bigcup_{b \in B} b} f \, d\mu .\]

The notation μ(B) is offered to compute $\mu (\bigcup_{b \in B} b)$.

GAIO.AdaptiveBoxMap — Method

BoxMap(:adaptive, f, domain::Box) -> SampledBoxMap

Construct a SampledBoxMap which uses sample_adaptive to generate test points as described in

Oliver Junge. “Rigorous discretization of subdivision techniques”. In: International Conference on Differential Equations. Ed. by B. Fiedler, K. Gröger, and J. Sprekels. 1999.

GAIO.GridBoxMap — Method

BoxMap(:grid, map, domain::Box{N,T}; no_of_points::NTuple{N} = ntuple(_->16, N)) -> SampledBoxMap
GridBoxMap(:grid, P::BoxPartition{N,T}; no_of_points::NTuple{N} = ntuple(_->16, N)) -> SampledBoxMap

Construct a SampledBoxMap that uses a grid of test points. The size of the grid is defined by no_of_points, which is a tuple of length equal to the dimension of the domain.

GAIO.MonteCarloBoxMap — Method

BoxMap(:montecarlo, map, domain::Box{N,T}; no_of_points=16*N) -> SampledBoxMap
BoxMap(:montecarlo, map, P::BoxPartition{N,T}; no_of_points=16*N) -> SampledBoxMap

Construct a SampledBoxMap that uses no_of_points Monte-Carlo test points.

GAIO.PointDiscretizedBoxMap — Method

BoxMap(:pointdiscretized, map, domain, points) -> SampledBoxMap

Construct a SampledBoxMap that uses the iterator points as test points. points must be an array or iterator of test points within the unit cube [-1,1]^N.

GAIO.PointDiscretizedBoxMap — Method

BoxMap(:pointdiscretized, :cpu, map, domain, points) -> CPUSampledBoxMap

Construct a CPUSampledBoxMap that uses the iterator points as test points. points must have eltype SVector{N, SIMD.Vec{S,T}} and be within the unit cube [-1,1]^N.

GAIO.PointDiscretizedBoxMap — Method

BoxMap(:pointdiscretized, :gpu, map, domain, points) -> GPUSampledBoxMap

Construct a GPUSampledBoxMap that uses the Vector points as test points. points must be a VECTOR of test points within the unit cube [-1,1]^N.

Requires a CUDA-capable gpu.

GAIO.adaptive_newton_step — Function

adaptive_newton_step(g, g_jacobian, x, k=1)

Return one step of the adaptive Newton algorithm for the point x.

The optional argument k is the iteration number, which is used to tune the step size.

GAIO.approx_lipschitz — Method

approx_lipschitz(f, center::SVector, radius::SVector, accel=nothing) -> Matrix

Compute an approximation of the Lipschitz matrix, i.e. the matrix that satisifies

\[| f(x) - f(y) | \leq L | x - y | \quad \forall \, x,y \in \text{Box(center, radius)}\]

componentwise.

GAIO.bounded_point_to_key — Method

bounded_point_to_key(partition::BoxPartition, point)

Find the cartesian index of the nearest box within a BoxPartition to a point. Conicides with point_to_key if the point lies in the partition. Default behavior is to set NaN = Inf if NaNs are present in point.

GAIO.center — Method

center(b::Box)
center(center, radius)

Return the center of a box as an iterable. Default function for image_points in SampledBoxMaps.

GAIO.chain_recurrent_set — Method

chain_recurrent_set(F::BoxMap, B::BoxSet; steps=12) -> BoxSet

Compute the chain recurrent set over the box set B. Generally, B should be a box set containing the whole partition P, ie B = P[:].

GAIO.cover — Method

BoxSet constructors:

set of all boxes in partition / box set P:

julia B = cover(P, :)

cover the point x, or points x = [x_1, x_2, x_3] # etc ... using boxes from P

julia B = cover(P, x)

a covering of S using boxes from P

julia S = [Box(center_1, radius_1), Box(center_2, radius_2), Box(center_3, radius_3)] # etc... B = cover(P, S)

Return a subset of the partition or box set P based on the second argument.

GAIO.cover_boxes — Method

cover_boxes(partition::BoxPartition, boxes)

Return a covering of an iterator of Boxes using Boxes from partition. Only covers the part of boxes which lies within partition.domain. This is returned by cover(partition, boxes).

GAIO.cover_roots — Method

cover_roots(g, Dg, B::BoxSet; steps=12) -> BoxSet

Compute a covering of the roots of g within the partition P. Generally, B should be a box set containing the whole partition P, ie B = P[:], and should contain a root of g.

GAIO.depth — Method

depth(tree::TreePartition)

Return the depth of the tree structure.

GAIO.find_at_depth — Method

find_at_depth(tree, depth)

Return all node indices at a specified depth.

GAIO.finite_time_lyapunov_exponents — Method

finite_time_lyapunov_exponents(F::BoxMap, boxset::BoxSet) -> BoxFun

Compute the Finite Time Lyapunov Exponent for every box in boxset, where F represents a time-T integration of some continuous dynamical system. It is assumed that all boxes in boxset have radii of some fixed order ϵ.

GAIO.hidden_keys — Method

hidden_keys(tree)

Return all keys within the tree, including keys not corresponding to leaf nodes.

GAIO.index_to_key — Method

index_to_key(iterable, i)

Return the object held in the ith position of iterable. Used to enumerate BoxSets as $\left{ B_1, B_2, \ldots, B_n \right}$ in TransferOperator, BoxGraph.

GAIO.key_to_box — Method

key_to_box(P::BoxPartition, key)

Return the box associated with the index within a BoxPartition.

GAIO.key_to_index — Method

key_to_index(iterable, key)

Find the index in 1..length(iterable) which holds key, or return nothing. Used to enumerate BoxSets as $\left{ B_1, B_2, \ldots, B_n \right}$ in TransferOperator, BoxGraph.

GAIO.leaves — Method

leaves(tree, initial_node_idx=1)

Return the node indices of all leaves. Begins search at initial_node_idx, i.e. only returns node indices of nodes below initial_node_idx within the tree.

GAIO.plotboxes! — Method

plot(boxset::BoxSet)
plot(boxfun::BoxFun)
plot!(boxset::BoxSet)
plot!(boxfun::BoxFun)

Plot a BoxSet or BoxFun.

Special Attributes:

projection = x -> x[1:3] If the dimension of the system is larger than 3, use this function to project to 3-d space.

color = :red Color used for the boxes.

colormap = :default Colormap used for plotting BoxFuns values.

marker = HyperRectangle(GeometryBasics.Vec3f0(0), GeometryBasics.Vec3f0(1)) The marker for an individual box. Only works if using Makie for plotting.

All other attributes are taken from MeshScatter.

GAIO.plotboxes — Method

plot(boxset::BoxSet)
plot(boxfun::BoxFun)
plot!(boxset::BoxSet)
plot!(boxfun::BoxFun)

Plot a BoxSet or BoxFun.

Special Attributes:

projection = x -> x[1:3] If the dimension of the system is larger than 3, use this function to project to 3-d space.

color = :red Color used for the boxes.

colormap = :default Colormap used for plotting BoxFuns values.

marker = HyperRectangle(GeometryBasics.Vec3f0(0), GeometryBasics.Vec3f0(1)) The marker for an individual box. Only works if using Makie for plotting.

All other attributes are taken from MeshScatter.

GAIO.point_to_box — Method

point_to_box(P::AbstractBoxPartition, point)

Find the box within a BoxPartition containing a point.

GAIO.point_to_key — Method

point_to_key(P::BoxPartition, point)

Find the index for the box within a BoxPartition contatining a point, or nothing if the point does not lie in the domain.

GAIO.relative_attractor — Method

relative_attractor(F::BoxMap, B::BoxSet; steps=12) -> BoxSet

Compute the attractor relative to B. Generally, B should be a box set containing the whole partition P, ie B = P[:].

GAIO.rescale — Method

rescale(center, radius, points)

Return an iterable which calls rescale(center, radius, point) for each point in points.

GAIO.rescale — Method

rescale(points)

Return a function

(center, radius) -> rescale(center, radius, points)

Used in domain_points for BoxMap, PointDiscretizedMap.

GAIO.rescale — Method

rescale(box, point::Union{<:StaticVector{N,T}, <:NTuple{N,T}})
rescale(center, radius, point::Union{<:StaticVector{N,T}, <:NTuple{N,T}})

Scale a point within the unit box $[-1, 1]^N$ to lie within box = Box(center, radius).

GAIO.rk4 — Method

rk4(f, x, τ)

Compute one step with step size τ of the classic fourth order Runge-Kutta method.

GAIO.rk4_flow_map — Function

rk4_flow_map(f, x, step_size=0.01, steps=20)

Perform steps steps of the classic Runge-Kutta fourth order method, with step size step_size.

GAIO.sample_adaptive — Method

sample_adaptive(f, center::SVector, radius::SVector, accel=nothing)

Create a grid of test points using the adaptive technique described in

Oliver Junge. “Rigorous discretization of subdivision techniques”. In: International Conference on Differential Equations. Ed. by B. Fiedler, K. Gröger, and J. Sprekels. 1999.

GAIO.subdivide! — Method

subdivide!(tree::TreePartition, key::keytype(tree)) -> TreePartition
subdivide!(tree::TreePartition, depth::Integer) -> TreePartition

subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, key) -> BoxSet
subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, depth) -> BoxSet

Subdivide a TreePartition at key. Dimension along which the node is subdivided depends on the depth of the node.

GAIO.subdivide — Method

subdivide(B::BoxSet{<:Any,<:Any,<:TreePartition}) -> BoxSet

Bisect every box in boxset along an axis, giving rise to a new partition of the domain, with double the amount of boxes. Axis along which to bisect depends on the depth of the nodes.

GAIO.unstable_set — Method

unstable_set(F::BoxMap, B::BoxSet) -> BoxSet

Compute the unstable set for a box set B. Generally, B should be a small box surrounding a fixed point of F. The partition should be fine enough, since no subdivision occurs in this algorithm.

GAIO.vertices — Method

vertices(box)
vertices(center, radius)

Return an iterator over the vertices of a box = Box(center, radius).