Library Reference

GAIO.BoxMap

GAIO.BoxSet

GAIO.Node

Arpack.eigs

Arpack.eigs

Arpack.eigs

Base.getindex

Base.sum

GAIO.AdaptiveBoxMap

GAIO.PointDiscretizedMap

GAIO.adaptive_newton_step

GAIO.approx_lipschitz

GAIO.bounded_point_to_cartesian

GAIO.center

GAIO.chain_recurrent_set

GAIO.cover_boxes

GAIO.cover_roots

GAIO.depth

GAIO.finite_time_lyapunov_exponents

GAIO.key_to_box

GAIO.plotboxes

GAIO.plotboxes!

GAIO.point_to_box

GAIO.point_to_key

GAIO.relative_attractor

GAIO.rescale

GAIO.rescale

GAIO.rescale

GAIO.rk4

GAIO.rk4_flow_map

GAIO.sample_adaptive

GAIO.subdivide

GAIO.subdivide!

GAIO.unsafe_point_to_cartesian

GAIO.unstable_set

GAIO.vertices

GAIO.BoxMap — Method

BoxMap(map, domain::Box{N,T}, accel=nothing; no_of_points=4*N*pick_vector_width(T)) -> SampledBoxMap
BoxMap(map, P::BoxPartition{N,T}, accel=nothing; no_of_points=4*N*pick_vector_width(T)) -> SampledBoxMap

Construct a SampledBoxMap which uses no_of_points Monte-Carlo test points.

GAIO.BoxSet — Method

BoxSet(boxgraph, graph_index_or_indices) -> BoxSet

Construct a BoxSet from some index / indices of vertices in a BoxGraph.

GAIO.Node — Type

Node structure used for TreePartitions

Fields:

left and right refer to indices w.r.t. trp.nodes and

trp.regular_partitions for a TreePartition trp.

.

Arpack.eigs — Function

eigs(gstar::TransferOperator [; kwargs...]) -> (d[, v], nconv)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Arpack.eigs — Function

eigs(gstar::TransferOperator [; kwargs...]) -> (d[, v], nconv)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Arpack.eigs — Function

eigs(gstar::TransferOperator [; kwargs...]) -> (d[, v], nconv)

Compute a set of eigenvalues d and eigenmeasures v of gstar. Works with the adjoint Koopman operator as well. All keyword arguments from Arpack.eigs can be passed. See the documentation for Arpack.eigs.

Base.getindex — Method

getindex constructors:

set of all boxes in P:

julia B = P[:]

cover the point x, or points x = [x_1, x_2, x_3] # etc ... using boxes from P

julia B = P[x]

a covering of S using boxes from P

julia S = [Box(center_1, radius_1), Box(center_2, radius_2), Box(center_3, radius_3)] # etc... B = P[S]

Return a subset of the partition P based on the second argument.

Base.sum — Method

sum(f, μ::BoxFun)
sum(f, μ::BoxFun, B::BoxSet)
μ(B) = sum(identity, μ, B)

Integrate a function f using μ as a density, that is, if boxfun is the discretization of a measure $\mu$ over the domain $Q$, then approximate the value of

\[\int_Q f \, d\mu .\]

If a BoxSet B is passed as the third argument, then the integration is restricted to the boxes in B

\[\int_{Q \cap \bigcup_{b \in B} b} f \, d\mu .\]

The notation μ(B) is offered to compute $\mu (\bigcup_{b \in B} b)$.

GAIO.AdaptiveBoxMap — Method

AdaptiveBoxMap(f, domain::Box, accel=nothing) -> SampledBoxMap

Construct a SampledBoxMap which uses sample_adaptive to generate test points.

GAIO.PointDiscretizedMap — Function

PointDiscretizedMap(map, domain, points, accel=nothing) -> SampledBoxMap

Construct a SampledBoxMap that uses the iterator points as test points. points must be an array or iterator of test points within the unit cube $[-1,1]^N$.

GAIO.adaptive_newton_step — Function

adaptive_newton_step(g, g_jacobian, x, k=1)

Return one step of the adaptive Newton algorithm for the point x.

The optional argument k is the iteration number, which is used to tune the step size.

GAIO.approx_lipschitz — Method

approx_lipschitz(f, center::SVector, radius::SVector, accel=nothing) -> Matrix

Compute an approximation of the Lipschitz matrix, i.e. the matrix that satisifies

\[| f(x) - f(y) | \leq L | x - y | \quad \forall \, x,y \in \text{Box(center, radius)}\]

componentwise.

GAIO.bounded_point_to_cartesian — Method

bounded_point_to_cartesian(partition::BoxPartition, point)

Find the cartesian index of the nearest box within a BoxPartition to a point. Conicides with unsafe_point_to_ints if the point lies in the partition.

GAIO.center — Method

center(b::Box)
center(center, radius)

Return the center of a box as an iterable. Default function for image_points in SampledBoxMaps.

GAIO.chain_recurrent_set — Method

chain_recurrent_set(F::BoxMap, B::BoxSet; steps=12) -> BoxSet

Compute the chain recurrent set over the box set B. Generally, B should be a box set containing the whole partition P, ie B = P[:].

GAIO.cover_boxes — Method

cover_boxes(partition::BoxPartition, boxes)

Return a covering of an iterator of Boxes using Boxes from partition. Only covers the part of boxes which lies within partition.domain.

GAIO.cover_roots — Method

cover_roots(g, Dg, B::BoxSet; steps=12) -> BoxSet

Compute a covering of the roots of g within the partition P. Generally, B should be a box set containing the whole partition P, ie B = P[:], and should contain a root of g.

GAIO.depth — Method

Return the depth of the tree structure.

GAIO.finite_time_lyapunov_exponents — Method

finite_time_lyapunov_exponents(F::BoxMap, boxset::BoxSet) -> BoxFun

Compute the Finite Time Lyapunov Exponent for every box in boxset, where F represents a time-T integration of some continuous dynamical system. It is assumed that all boxes in boxset have radii of some fixed order ϵ.

GAIO.key_to_box — Method

key_to_box(P::BoxPartition, key)

Return the box associated with the index within a BoxPartition.

GAIO.plotboxes! — Method

plot(boxset::BoxSet)
plot(boxfun::BoxFun)
plot!(boxset::BoxSet)
plot!(boxfun::BoxFun)

Plot a BoxSet or BoxFun.

Special Attributes:

projection = x -> x[1:3] If the dimension of the system is larger than 3, use this function to project to 3-d space.

color = :red Color used for the boxes.

colormap = :default Colormap used for plotting BoxFuns values.

marker = HyperRectangle(GeometryBasics.Vec3f0(0), GeometryBasics.Vec3f0(1)) The marker for an individual box. Only works if using Makie for plotting.

All other attributes are taken from MeshScatter.

GAIO.plotboxes — Method

plot(boxset::BoxSet)
plot(boxfun::BoxFun)
plot!(boxset::BoxSet)
plot!(boxfun::BoxFun)

Plot a BoxSet or BoxFun.

Special Attributes:

projection = x -> x[1:3] If the dimension of the system is larger than 3, use this function to project to 3-d space.

color = :red Color used for the boxes.

colormap = :default Colormap used for plotting BoxFuns values.

marker = HyperRectangle(GeometryBasics.Vec3f0(0), GeometryBasics.Vec3f0(1)) The marker for an individual box. Only works if using Makie for plotting.

All other attributes are taken from MeshScatter.

GAIO.point_to_box — Method

point_to_box(P::BoxPartition, point)

Find the box within a BoxPartition containing a point.

GAIO.point_to_key — Method

point_to_key(P::BoxPartition, point)

Find the index for the box within a BoxPartition contatining a point.

Bounds checking

unlike unsafe_point_to_ints, point_to_key will return nothing if the point does not lie in the partition.

GAIO.relative_attractor — Method

relative_attractor(F::BoxMap, B::BoxSet; steps=12) -> BoxSet

Compute the attractor relative to B. Generally, B should be a box set containing the whole partition P, ie B = P[:].

GAIO.rescale — Method

rescale(center, radius, points)

Return an iterable which calls rescale(center, radius, point) for each point in points.

GAIO.rescale — Method

rescale(points)

Return a function

(center, radius) -> rescale(center, radius, points)

Used in domain_points for BoxMap, PointDiscretizedMap.

GAIO.rescale — Method

rescale(box, point::Union{<:StaticVector{N,T}, <:NTuple{N,T}})
rescale(center, radius, point::Union{<:StaticVector{N,T}, <:NTuple{N,T}})

Scale a point within the unit box $[-1, 1]^N$ to lie within box = Box(center, radius).

GAIO.rk4 — Method

rk4(f, x, τ)

Compute one step with step size τ of the classic fourth order Runge-Kutta method.

GAIO.rk4_flow_map — Function

rk4_flow_map(f, x, step_size=0.01, steps=20)

Perform steps steps of the classic Runge-Kutta fourth order method, with step size step_size.

GAIO.sample_adaptive — Method

sample_adaptive(f, center::SVector, radius::SVector, accel=nothing)

Create a grid of test points using the adaptive technique described in

Oliver Junge. “Rigorous discretization of subdivision techniques”. In: International Conference on Differential Equations. Ed. by B. Fiedler, K. Gröger, and J. Sprekels. 1999.

GAIO.subdivide! — Method

subdivide!(tree:TreePartition), key::NTuple{2,<:Integer}) -> TreePartition
subdivide!(boxset::BoxSet{<:Any,<:Any,<:TreePartition}, key::NTuple{2,<:Integer}) -> BoxSet

Subdivide a TreePartition at the node key. Dimension along which the node is subdivided depends on the depth of the node.

GAIO.subdivide — Method

subdivide(B::BoxSet{<:Any,<:Any,<:TreePartition}) -> BoxSet

Bisect every box in boxset along an axis, giving rise to a new partition of the domain, with double the amount of boxes. Axis along which to bisect depends on the depth of the nodes.

GAIO.unsafe_point_to_cartesian — Method

unsafe_point_to_cartesian(P::BoxPartiton, point)

Find the cartesian index for the box within a BoxPartition containing a point.

bounds checking

unsafe_point_to_ints does not do any bounds checking. The returned cartesian index will be out of bounds if the point does not lie in the partition.

GAIO.unstable_set — Method

unstable_set(F::BoxMap, B::BoxSet) -> BoxSet

Compute the unstable set for a box set B. Generally, B should be a small box surrounding a fixed point of F. The partition should be fine enough, since no subdivision occurs in this algorithm.

GAIO.vertices — Method

vertices(box)
vertices(center, radius)

Return an iterator over the vertices of a box = Box(center, radius).